% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_bivariado_multivariado.R
\name{est_sig}
\alias{est_sig}
\title{Estimate Sigma in Multivariate Case}
\usage{
est_sig(data, exx, ex, updmu)
}
\arguments{
\item{data}{Data in the form of multivariate grouped data.}

\item{exx}{The list of second moments matrices.}

\item{ex}{Matrix with the references of the first moments.}

\item{updmu}{The estimate mu.}
}
\value{
returns a matrix with the variance and covariance estimates.
}
\description{
The function to calculate the estimate of the covariances,
also for multivariate case.
}
\examples{

library(MASS)
library(em.univ)
library(tmvtnorm)

simulateddata = em.univ::mult_simul(mm = c(68,68),
                          ss = base::matrix(c(3,2,2,6),2,2) ,
                          n_data_sets = 1,
                          breaks_x = c(-Inf,64,65,66,67,68,69,70,71,72,Inf),
                          breaks_y = c(-Inf,64.2,65.2,66.2,
                                       67.2,68.2,69.2,70.2,
                                       71.2,72.2,Inf),
                          lower_x = base::rep(c(-Inf,64,65,66,67,
                                          68,69,70,71,72),10),
                          lower_y = c(base::rep(-Inf,10),
                                      base::rep(64.2,10),
                                      base::rep(65.2,10),
                                      base::rep(66.2,10),
                                      base::rep(67.2,10),
                                      base::rep(68.2,10),
                                      base::rep(69.2,10),
                                      base::rep(70.2,10),
                                      base::rep(71.2,10),
                                      base::rep(72.2,10)),
                          upper_x = base::rep(c(64,65,66,67,68,69,70,
                                                71,72,Inf),10),
                          upper_y = c(base::rep(64.2,10),
                                      base::rep(65.2,10),
                                      base::rep(66.2,10),
                                      base::rep(67.2,10),
                                      base::rep(68.2,10),
                                      base::rep(69.2,10),
                                      base::rep(70.2,10),
                                      base::rep(71.2,10),
                                      base::rep(72,2,10),
                                      base::rep(Inf,10))
)

mu2<- c(67,67)
sigma2<- base::matrix(c(3.1,2.2,2.2,4.3),2,2)


exi<- em.univ::mexi(data=simulateddata[,,1],
           mu=mu2,
           sigma=sigma2)

ss1<- em.univ::mcovxi(data=simulateddata[,,1],
             mu=mu2,
             sigma=sigma2)

exx<- em.univ::exxest(data=simulateddata[,,1],
             exi=exi,
             ss1=ss1)

updmu <- em.univ::mem(data=simulateddata[,,1],
             exi=exi)

out<- em.univ::est_sig(data=simulateddata[,,1],
            exx=exx,
            ex=exi,
            updmu = updmu)


out

}
