% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_bivariado_multivariado.R
\name{mcem_mult}
\alias{mcem_mult}
\title{Estimate mean and Variance by MCEM multivariate}
\usage{
mcem_mult(data, mu_init, sigma_init, maxit = 1000, tol1 = 0.001, tol2 = 0.001)
}
\arguments{
\item{data}{Data in the form of multivariated grouped data.}

\item{mu_init}{The initialized mean vector.}

\item{sigma_init}{The initialize covariance matrix.}

\item{maxit}{Integer, number of max iterations .}

\item{tol1}{The stopping criterias for means vectors.}

\item{tol2}{The stopping criterias for covariance matrix.}
}
\value{
Returns a list with the mu estimate and the variance matrix.
}
\description{
This is the function of M-step, to iterate over the results
for mean and variance.
}
\examples{

library(MASS)
library(em.univ)
library(tmvtnorm)

simulateddata = em.univ::mult_simul(mm = c(68,68),
                          ss = base::matrix(c(3,2,2,6),2,2) ,
                          n_data_sets = 1,
                          breaks_x = c(-Inf,64,65,66,67,68,69,70,71,72,Inf),
                          breaks_y = c(-Inf,64.2,65.2,66.2,
                                       67.2,68.2,69.2,70.2,
                                       71.2,72.2,Inf),
                          lower_x = base::rep(c(-Inf,64,65,66,67,
                                          68,69,70,71,72),10),
                          lower_y = c(base::rep(-Inf,10),
                                      base::rep(64.2,10),
                                      base::rep(65.2,10),
                                      base::rep(66.2,10),
                                      base::rep(67.2,10),
                                      base::rep(68.2,10),
                                      base::rep(69.2,10),
                                      base::rep(70.2,10),
                                      base::rep(71.2,10),
                                      base::rep(72.2,10)),
                          upper_x = base::rep(c(64,65,66,67,68,69,70,71,72,Inf),10),
                          upper_y = c(base::rep(64.2,10),
                                      base::rep(65.2,10),
                                      base::rep(66.2,10),
                                      base::rep(67.2,10),
                                      base::rep(68.2,10),
                                      base::rep(69.2,10),
                                      base::rep(70.2,10),
                                      base::rep(71.2,10),
                                      base::rep(72,2,10),
                                      base::rep(Inf,10))
)

mu2<- c(67,67)
sigma2<- base::matrix(c(3.1,2.2,2.2,4.3),2,2)


out<- em.univ::mcem_mult(data=simulateddata[,,1],
              mu_init=mu2,
              sigma_init=sigma2,
              maxit=1000,
              tol1=1e-4,
              tol2=1e-3)

out

}
