% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_mcem_logll.R
\name{mest}
\alias{mest}
\title{E-Step For Mean Estimate}
\usage{
mest(theta, bl, bu, freq)
}
\arguments{
\item{theta}{Initial theta value, vector containing mean and sd.}

\item{bl}{Lower bound of the intervals, values in vector, starting from -inf.}

\item{bu}{Upper bound of the intervals, values in vector, ending with +inf.}

\item{freq}{frequency over the intervals, values in vector.}
}
\value{
Return M which are the estimates of mean in E-step.
}
\description{
E-Step For Mean Estimate
}
\examples{

 library(em.univ)
 library(stats)
 
 simdataaaa <- em.univ::univ_simul(ncol_matrix=1,
                         n=50,
                         nclass = 10,
                         mean = 68,
                         sd = 1.80,
                         fr_breaks=c(62,64,66,68,70,72,74,76,78))
 
 munew <- em.univ::mest(theta=c(67,2),
               bl=simdataaaa$simul_data[,1,1],
               bu=simdataaaa$simul_data[,2,1],
               freq=simdataaaa$simul_data[,3,1])
 
 
 munew
}
