% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_bivariado_multivariado.R
\name{mult_llik}
\alias{mult_llik}
\title{Defining the Log-Likelihood Function}
\usage{
mult_llik(data, theta)
}
\arguments{
\item{data}{Data in the form of multivariated grouped data.}

\item{theta}{Is a the vector of all parameters, the first d elements
of it are the means, d+1 are v varinces and d+v+1 are c covariances.
In the form:
theta=[mu_1,...,mu_d,
Var_1,...,Var_v,
Cov_1(1,(1+1)), ...,Cov_d(1,d),
Cov_d+1(2,(2+1)), ...,Cov_c(d-1,d)].}
}
\value{
Return the optimum piont or the maximum to log-likelihood
}
\description{
In this function we will calculate the log-likelihood function.
}
\examples{

library(MASS)
library(em.univ)
library(tmvtnorm)

simulateddata = em.univ::mult_simul(mm = c(68,68),
                          ss = base::matrix(c(3,2,2,6),2,2) ,
                          n_data_sets = 1,
                          breaks_x = c(-Inf,64,65,66,67,68,69,70,71,72,Inf),
                          breaks_y = c(-Inf,64.2,65.2,66.2,
                                       67.2,68.2,69.2,70.2,
                                       71.2,72.2,Inf),
                          lower_x = base::rep(c(-Inf,64,65,66,67,
                                          68,69,70,71,72),10),
                          lower_y = c(base::rep(-Inf,10),
                                      base::rep(64.2,10),
                                      base::rep(65.2,10),
                                      base::rep(66.2,10),
                                      base::rep(67.2,10),
                                      base::rep(68.2,10),
                                      base::rep(69.2,10),
                                      base::rep(70.2,10),
                                      base::rep(71.2,10),
                                      base::rep(72.2,10)),
                          upper_x = base::rep(c(64,65,66,67,68,69,70,
                                                71,72,Inf),10),
                          upper_y = c(base::rep(64.2,10),
                                      base::rep(65.2,10),
                                      base::rep(66.2,10),
                                      base::rep(67.2,10),
                                      base::rep(68.2,10),
                                      base::rep(69.2,10),
                                      base::rep(70.2,10),
                                      base::rep(71.2,10),
                                      base::rep(72,2,10),
                                      base::rep(Inf,10))
)

theta<- c(65,65,3,6,2.12132)

out<- em.univ::mult_llik(data = simulateddata[,,1],
                theta = theta)

out

}
