% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_mcem_logll.R
\name{mumcem}
\alias{mumcem}
\title{E-Step estimate the parameter of mu.}
\usage{
mumcem(data, simz)
}
\arguments{
\item{data}{Contingency table, matrix format of three columns, first column
lower limits, second column upper limits, and third column observed
frequencies. The current state of the parameters, vector containing
mean and sd.}

\item{simz}{The matrix, which is the samples simulated over the intervals,
of the zmcem function.}
}
\value{
Return mu which are the estimates of mean in E-step.
}
\description{
E-Step estimate the parameter of mu.
}
\examples{

 simdataaaa <- em.univ::univ_simul(ncol_matrix=1,
                         n=50,
                         nclass = 10,
                         mean = 68,
                         sd = 1.80,
                         fr_breaks=c(62,64,66,68,70,72,74,76,78))
 
 zm_matriz <- em.univ::zmcem(theta=c(67,2),data = simdataaaa$simul_data[,,1])
 
 mu_estimate <- em.univ::mumcem(data = simdataaaa$simul_data[,,1],
                           simz = zm_matriz)
 mu_estimate
}
