% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_mcem_logll.R
\name{ssest}
\alias{ssest}
\title{E-Step For Variance Estimate}
\usage{
ssest(theta, bl, bu, muupdate, freq)
}
\arguments{
\item{theta}{Initial theta value, vector containing mean and sd.}

\item{bl}{Lower bound of the intervals, values in vector, starting from -inf.}

\item{bu}{Upper bound of the intervals, values in vector, ending with +inf.}

\item{muupdate}{Is the updated estimates of mu.}

\item{freq}{Frequency over the intervals, values in vector.}
}
\value{
Return ss which are the estimates of sigma (variance) in E-step.
}
\description{
E-Step For Variance Estimate
}
\examples{

 library(em.univ)
 library(stats)


 simdataaaa <- em.univ::univ_simul(ncol_matrix=1,
                        n=50,
                        nclass = 10,
                        mean = 68,
                        sd = 1.80,
                        fr_breaks=c(62,64,66,68,70,72,74,76,78))
 
 ssnew <- em.univ::ssest(theta=c(67,2),
              bl=simdataaaa$simul_data[,1,1],
              bu=simdataaaa$simul_data[,2,1],
              muupdate=mest(theta=c(67,2),
                            bl=simdataaaa$simul_data[,1,1],
                            bu=simdataaaa$simul_data[,2,1],
                            freq=simdataaaa$simul_data[,3,1]),
              freq=simdataaaa$simul_data[,3,1])
 
 ssnew
}
