% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_mcem_logll.R
\name{univ_simul}
\alias{univ_simul}
\title{Generating data from a gaussian}
\usage{
univ_simul(
  ncol_matrix = 30,
  n = 50,
  nclass = 10,
  mean = 68,
  sd = 1.8,
  fr_breaks = c(62, 64, 66, 68, 70, 72, 74, 76, 78)
)
}
\arguments{
\item{ncol_matrix}{A number, number of repetitions of the random simulated
samples.}

\item{n}{A number, number of observations in each sample.}

\item{nclass}{A number, number of classes on the contingency table.}

\item{mean}{A number, parameter mu value to generate data.}

\item{sd}{A number, parameter sigma value to generate data.}

\item{fr_breaks}{A vector, vector containing the values of the referenced
as counts.}
}
\value{
Returns a list with two objects, the first "simul_data" returns
 a matrix in which the first column the lower limits of the intervals,
 in the second column the values of the upper limits of the intervals and
 the third column the counts of the observations. The second object "med"
 returns a matrix, in which the first column shows the sampled value and
 the second column the observation count.
}
\description{
Generating data from a gaussian
}
\examples{

 library(em.univ)
 library(stats)
 
 simdataaaa <- em.univ::univ_simul(ncol_matrix=1,
                       n=50,
                       nclass = 10,
                       mean = 68,
                       sd = 1.80,
                       fr_breaks=c(62,64,66,68,70,72,74,76,78))
                       
 simdataaaa
 
}
